inserexs file and directories
The reading of the user manual is strongly recommended, as well as the license note. The list of needed packages is fully indicated.
The "main.py" file acts as backbone for the program execution. Running it lunches the graphic interface.
The FDMNES software should be downloaded and copied into the "FDMNES" file.
inserexs is free to use and modify with a proper authorship recognition.